+++ zera-basemodule/intvalidator.h
@@ -6,15 +6,15 @@
 class cIntValidator: public cParamValidator
 {
 public:
-    cIntValidator(qint32 min, qint32 max, qint32 delta = 0);
+    cIntValidator(qint64 min, qint64 max, qint64 delta = 0);
 
     virtual bool isValidParam(QVariant newValue);
     virtual void exportMetaData(QJsonObject& jsObj);
 
 private:
-    qint32 m_nMin;
-    qint32 m_nMax;
-    qint32 m_nDelta;
+    qint64 m_nMin;
+    qint64 m_nMax;
+    qint64 m_nDelta;
 };
 
 #endif // INTVALIDATOR_H
